# backend/doc/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "High Street Gym API",
    "version": "1.0.0",
    "description": "JSON REST API for interacting with the High Street Gym Backend"
  },
  "servers": [
    {
      "url": "http://localhost:8080/api",
      "description": "Local development server"
    }
  ],
  "components": {
    "securitySchemes": {
      "apiKey": {
        "type": "apiKey",
        "in": "header",
        "name": "x-auth-key",
        "description": "API key for authentication"
      }
    },
    "schemas": {
...
      "Error": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "description": "Error message",
            "example": "An error occurred"
          }
        }
      },
...
      "Booking": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "description": "The unique identifier of the booking",
            "example": 1
          },
          "memberId": {
            "type": "number",
            "description": "The ID of the member who made the booking",
            "example": 2
          },
          "sessionId": {
            "type": "number",
            "description": "The ID of the session being booked",
            "example": 5
          },
          "deleted": {
            "type": "number",
            "description": "Deletion status (0 = active, 1 = deleted)",
            "example": 0
          }
        }
      },
      "BookingInput": {
        "type": "object",
        "required": [
          "sessionId"
        ],
        "properties": {
          "sessionId": {
            "type": "number",
            "description": "The ID of the session to book",
            "example": 5
          }
        },
        "example": {
          "sessionId": 5
        }
      }
    },
    "responses": {
...
      "BadRequest": {
        "description": "Invalid request body or bad request",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "Unauthorized": {
        "description": "Authentication required or invalid credentials",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "NotFound": {
        "description": "Resource not found",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "Conflict": {
        "description": "Conflict - resource already exists",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            },
            "example": {
              "message": "You are already booked for this session"
            }
          }
        }
      },
      "Forbidden": {
        "description": "Access forbidden - insufficient permissions",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "InternalServerError": {
        "description": "Server error",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "SuccessMessage": {
        "description": "Success response with a message",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "description": "Success message",
                  "example": "Operation successful"
                }
              }
            }
          }
        }
      },
...
      "BookingCreated": {
        "description": "Booking created successfully",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Booking"
            }
          }
        }
      },
      "BookingsList": {
        "description": "List of bookings",
        "content": {
          "application/json": {
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/components/schemas/Booking"
              }
            }
          }
        }
      }
    }
  },
  "paths": {
...
    "/bookings/self": {
      "get": {
        "summary": "Get authenticated member's bookings",
        "tags": [
          "Bookings"
        ],
        "description": "Retrieve a list of all active bookings for the authenticated member",
        "security": [
          {
            "apiKey": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "includePast",
            "schema": {
              "type": "string",
              "enum": [
                "true",
                "false"
              ]
            },
            "description": "If 'true', includes past bookings in the results. Default is 'false' (only future bookings).",
            "required": false
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/BookingsList"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      }
    },
    "/bookings/{id}": {
      "get": {
        "summary": "Get a specific booking",
        "tags": [
          "Bookings"
        ],
        "description": "Retrieve a specific booking with full session, activity, location, and trainer details. Members can only view their own bookings.",
        "security": [
          {
            "apiKey": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "The booking ID"
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/BookingCreated"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      },
      "delete": {
        "summary": "Cancel a booking",
        "tags": [
          "Bookings"
        ],
        "description": "Cancel a booking by ID. Only the member who owns the booking can cancel it.",
        "security": [
          {
            "apiKey": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "The booking ID"
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/SuccessMessage"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      }
    },
    "/bookings": {
      "post": {
        "summary": "Create a new booking",
        "tags": [
          "Bookings"
        ],
        "description": "Create a booking for the authenticated member by selecting a session. Members can only create bookings for themselves.",
        "security": [
          {
            "apiKey": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BookingInput"
              }
            }
          }
        },
        "responses": {
          "201": {
            "$ref": "#/components/responses/BookingCreated"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      }
    },
    "/bookings/export/xml/history": {
      "get": {
        "summary": "Export member's booking history as XML",
        "tags": [
          "Bookings"
        ],
        "description": "Export the authenticated member's booking history as XML for fitness tracking applications. Includes activity data, session details, location, and trainer information.",
        "security": [
          {
            "apiKey": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "onlyPast",
            "schema": {
              "type": "string",
              "enum": [
                "true",
                "false"
              ]
            },
            "description": "If 'true', exports only previous (past) bookings. Default is 'false' (all bookings).",
            "required": false
          }
        ],
        "responses": {
          "200": {
            "description": "XML file containing member's booking history",
            "content": {
              "application/xml": {
                "schema": {
                  "type": "string",
                  "example": "<?xml version=\"1.0\" encoding=\"UTF-8\"?>\n<booking_history>\n  <header>\n    <title>Booking History - John Doe</title>\n    <member>\n      <name>John Doe</name>\n      <email>member@hsg.com</email>\n    </member>\n  </header>\n  <bookings>...</bookings>\n</booking_history>\n"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      }
    },
...
  },
...
}